requirements.txt
contourpy==1.2.0
cycler==0.12.1
fonttools==4.47.0
imageio==2.33.1
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.2
networkx==3.2.1
numpy==1.26.3
packaging==23.2
pillow==10.2.0
PyMaxflow==1.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
threadpoolctl==3.2.0
tifffile==2023.12.9
